@@ -34,11 +34,11 @@
     "coveralls": "^2.11.12",
     "css-loader": "^0.26.1",
     "enzyme": "^2.4.1",
-    "eslint": "^3.7.1",
-    "eslint-config-airbnb": "^13.0.0",
+    "eslint": "^3.13.1",
+    "eslint-config-airbnb": "^14.0.0",
     "eslint-plugin-import": "^2.2.0",
-    "eslint-plugin-jsx-a11y": "^2.2.3",
-    "eslint-plugin-react": "^6.4.0",
+    "eslint-plugin-jsx-a11y": "^3.0.2",
+    "eslint-plugin-react": "^6.9.0",
     "glob": "^7.1.1",
     "inject-loader": "^2.0.1",
     "jest": "^18.1.0",
